+++ packages/plugin-access/src/utils/getAccessProviderContent.ts
@@ -33,12 +33,13 @@ const AccessProvider: React.FC<Props> = props => {
     console.warn('[plugin-access]: the access instance created by access.ts(js) is nullish, maybe you need check it.');
   }
 
-  props.routes.splice(0, props.routes.length, ...traverseModifyRoutes(props.routes, access));
-
   return React.createElement(
     AccessContext.Provider,
     { value: access },
-    children,
+    React.cloneElement(children, {
+      ...children.props,
+      routes:traverseModifyRoutes(props.routes, access)
+    }),
   );
 };