+++ packages/plugin-request/tests/index.test.js
@@ -24,6 +24,7 @@ describe('plugin-request', () => {
           return '/winpathtest';
         },
       },
+      chainWebpack() {},
       addUmiExports() {},
       addDepInfo() {},
       writeTmpFile,